requirements.txt
Django==2.1.1
djangorestframework==3.9.2
Pillow==7.0.0
postgres==2.2.2
psycopg2-binary==2.7.7
django-cors-headers==2.5.0
requests==2.23.0
beautifulsoup4==4.8.2
APScheduler==3.6.0
Celery==4.4.0